CatCommand working: it now takes flag.args, not the args interface.
I had been using the array of interfaces directly as I thought
that was the best way to get to the arguments to the cat command.
However, that turns out to be an array of arrays of interface
and the implementation was getting really messy.

After a little thought and research I found that flags already
contained the parsed arguments to the command as an array of string
so that was the obvious solution and simplified the implementation.