Adds the template render command up to arg parsing.
This involved several large refactorings:
- the test helped no longer tracks things in an object-like structure. It's
just very simple structs and functions that operate on the type.

- the render command is an entirely new subcommand with a new type of argument:
a dynamically parsed argument.

You can see the parsing code in main.go; it extracts and sets up a map of
flags and their values, so we can pass yaml files in as template configs
containing the data that hydrates a template file.

- adds a render test file, but does not exercise the functionality yet.