Adds the template rendering functionality.
- Introduces two new types. This was a cool learning opportunity,
as the introduction of a templatePath and TemplateVariablesPath suddenly
forced the code to become much clearer about its intentions. Prior to
this, I had been using maps of strings and there was no information
in the type system about the code's intention.

- NewRenderCommand actually now returns a copy of the default rendercommand.
The functionality is precisely what I need today, with the templateconfigpaths
populated along with all the help messages being correct.

- Changed the default log level to warn after experimentation.