Implements the first render command test
It is a very hard method to test as it only supports a generic
return value presently. I can refactor this to get richer information
internally, while presenting the right public interface.

I added a new logging pattern: always adding an else clause
for every error check, so that the success case is also
acknowledged.

Getting the tests setup working was done by changing the test
data type from previous test setup work, to support the
newly required flexibility. That touched every test file.

Having the type names present helped with reasoning a couple of
times, so that was a win.