partial match (#4)
* updates catcommand to use a contains match for its file arguments.

This simplifies findFilesByNames to use a slice of strings, which
is a nice simplification.

* Finishes the rewrite of renders Execute method.

Fundamentally the issue with the old implemented of Execute and
render was that they were trying too hard. By simplifying the path
logic (note that template variable files were written to be absolute
paths starting in the config directory, which is changed now), and
making the findFilesByName function easier to interact with, I can
see real progress now.