refactoring bigly
- removes the dependency on subcommands. It's too heavyweight; this
tool can exist using the standard library to manage arguments.

- reimplements the repository functionality using simple  combinations
of types and an interface. There are constructors to return correctly
initialised instances of two repository types. First, the github repository,
and second it now excitingly supports local git repositories. This
really makes types of repository much more extensible; the old implementation
really was going to be hard to add new versions of.

- removes the cmds subdirectory. There are portions of the old
implementation of the logic that I dig and will resurrect, but it was not
very go-ish.

- configelements is a package to track configuration elements. I
can make it more generic, but it just tracks an environment variable
name to store config elements, a default and a way to set the value
to something new.

- pretty decent test coverage! The first version really hasn't ever
really had this.