Adds support back in for --fetching the repository.
I renamed the integration test script to be more specific while I
was in here.

Adding in git-go support brings in a lot of dependencies. One is
always tempted to think one could reimplement with far fewer
dependencies, but this is one where I absolutely respect the upstream
developers. git is big and tough and this library is a
pleasure to work with.